requirements.txt
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorlog==6.8.2
cryptography==44.0.2
Django==5.1.7
django-filter==24.2
django-jinja==2.11.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
idna==3.8
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
numpy==2.1.0
pandas==2.2.2
paramiko==3.4.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
